requirements.txt
tensorflow
streamlit
gtts
Pillow
